Conflicts:
	.gitignore
	Makefile.am
	beamline/include/SeptumPropagators.h
	beamline/include/septum.h
	beamline/src/CF_rbend.cc
	beamline/src/CF_sbend.cc
	beamline/src/SeptumPropagators.cc
	beamline/src/ThinPolesPropagators.cc
	beamline/src/beamline.cc
	beamline/src/bmlnElmnt.cc
	bmlfactory/src/MAD8Factory.cc
	bmlfactory/src/mad_yacc.y
	configure.in
	mxyzptlk/include/TJLterm.h
	mxyzptlk/include/TMapping.tcc
	physics_toolkit/src/CMakeLists.txt
	physics_toolkit/src/ChromaticityAdjuster.cc
	physics_toolkit/src/FPSolver.cc
	physics_toolkit/src/LattFuncSage.cc
	physics_toolkit/src/LayoutVisitor.cc
	python-bindings/src/CMakeLists.txt
	python-bindings/src/Makefile.am
	python-bindings/src/py-sextupole.cpp
	updateheaders
	updatesource

+++ beamline/src/bmlnElmnt.cc
@@ -576,7 +576,7 @@ void bmlnElmnt::peekAt( double& s, const Particle& prt ) const
 
  (*pcout) << setw(12) << s
                   << " : "
-      << setw(10) << this
+      << setw(10) << (long) this
       << setw(15) << ident_
       << setw(15) << Type()
       << setw(12) << length_
